Rename files to better fit with Arduino IDE
