+++ IrForLedStrip.h
